feat: make WeaponSet loader extensible
