--- src/models/Weapons.ts
@@ -1,18 +1,0 @@
-import WeaponSet from './WeaponSet';
-
-const Weapons: WeaponSet[] = [
-  {
-    name: 'ブキ1',
-    slug: 'weapon1',
-  },
-  {
-    name: 'ブキ2',
-    slug: 'weapon2',
-  },
-  {
-    name: 'ブキ3',
-    slug: 'weapon3',
-  },
-];
-
-export default Weapons;
