+++ src/repositories/WeaponSetRepository.ts
@@ -1,0 +1,7 @@
+import WeaponSet from '../models/WeaponSet';
+
+abstract class WeaponSetRepository {
+  abstract Get(): WeaponSet[];
+}
+
+export default WeaponSetRepository;
